requirements.txt
torch>=2.1.0
transformers>=4.40.0
wtpsplit>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0
pyyaml>=6.0
tqdm>=4.65.0
nltk>=3.8.0
spacy>=3.7.0
accelerate>=0.28.0
